+++ app/Http/Controllers/RwController.php
@@ -5,14 +5,16 @@ namespace App\Http\Controllers;
 use App\Http\Requests\GeneralSearchRequest;
 use App\Http\Requests\Rw\StoreRequest;
 use App\Interfaces\Controllers\HasSearch;
-use App\Interfaces\Services\Group\GroupServiceInterface;
+use App\Interfaces\Services\Tenant\TenantServiceInterface;
+use App\Models\Tenant;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 use Inertia\Response;
 
 class RwController extends Controller implements HasSearch
 {
-    public function __construct(protected GroupServiceInterface $service)
+    public function __construct(protected TenantServiceInterface $service)
     {
         parent::__construct();
     }
@@ -30,13 +32,15 @@ class RwController extends Controller implements HasSearch
      */
     public function index(GeneralSearchRequest $request): Response
     {
+        Gate::authorize('viewAnyRw', Tenant::class);
+
         $datas = $this->service->findAllPaginate(
             $this->per_page,
             fn($q) => $q->whereNull('parent_id')
                 ->with('leader')
         );
 
-        return Inertia::render('rw/index', [
+        return Inertia::render('rw/index/index', [
             'datas' => $datas,
             'filters' => [
                 'search' => $request->filter['search'] ?? ""
@@ -51,7 +55,9 @@ class RwController extends Controller implements HasSearch
      */
     public function create(): Response
     {
-        return Inertia::render('rw/create');
+        Gate::authorize('createRw', Tenant::class);
+
+        return Inertia::render('rw/create/index');
     }
 
     /**
@@ -59,6 +65,8 @@ class RwController extends Controller implements HasSearch
      */
     public function store(StoreRequest $request)
     {
+        Gate::authorize('createRw', Tenant::class);
+
         $this->service->create($request->validated());
         return to_route('rw.index')->with('success', self::CREATED_MESSAGE);
     }
@@ -68,14 +76,16 @@ class RwController extends Controller implements HasSearch
      */
     public function show(string $id): Response
     {
-        $group = $this->service->findById($id, [
+        Gate::authorize('viewRw', Tenant::class);
+
+        $tenant = $this->service->findById($id, [
             'leader' => fn($q) => $q->selectMinimalist(),
             'createdBy' => fn($q) => $q->selectMinimalist(),
             'updatedBy' => fn($q) => $q->selectMinimalist(),
         ]);
 
-        return Inertia::render('rw/show', [
-            'data' => $group
+        return Inertia::render('rw/show/index', [
+            'data' => $tenant
         ]);
     }
 
@@ -84,9 +94,11 @@ class RwController extends Controller implements HasSearch
      */
     public function edit(string $id)
     {
-        $group = $this->service->findById($id);
-        return Inertia::render('rw/edit', [
-            'data' => $group
+        Gate::authorize('updateRw', Tenant::class);
+
+        $tenant = $this->service->findById($id);
+        return Inertia::render('rw/edit/index', [
+            'data' => $tenant
         ]);
     }
 
@@ -95,6 +107,8 @@ class RwController extends Controller implements HasSearch
      */
     public function update(StoreRequest $request, string $id)
     {
+        Gate::authorize('updateRw', Tenant::class);
+
         $this->service->update($id, $request->validated());
         return to_route('rw.index')->with('success', self::UPDATED_MESSAGE);
     }
@@ -104,6 +118,8 @@ class RwController extends Controller implements HasSearch
      */
     public function destroy(string $id)
     {
+        Gate::authorize('deleteRw', Tenant::class);
+
         $this->service->delete($id);
         return redirect()->back()->with('success', self::DELETED_MESSAGE);
     }
